feat(GA) : seperate GA file
Refs:[#10]
